Move demo dataset definitions from MediaType into demo importer package
Demo dataset constants (category lists, DemoDataset objects, required_folder
paths) were scattered across each MediaType subclass but only consumed by
the demo importer pipeline. This made them impossible to remove as a unit.

Centralizes all definitions in vtsearch/datasets/importers/demo/datasets.py:
- DemoDataset dataclass moved from media/base.py
- Category constants moved from audio/image/text/video media_type.py
- all_demo_datasets() moved from media/__init__.py
- Removes abstract demo_datasets property from MediaType

load_demo_source() stays on MediaType since it contains media-type-specific
download and embedding logic.
