+++ vtsearch/media/document/media_type.py
@@ -82,14 +82,6 @@ class DocumentMediaType(MediaType):
         return False
 
     # ------------------------------------------------------------------
-    # Demo datasets
-    # ------------------------------------------------------------------
-
-    @property
-    def demo_datasets(self) -> list:
-        return []
-
-    # ------------------------------------------------------------------
     # Embeddings (not supported)
     # ------------------------------------------------------------------
