+++ vtsearch/media/image/media_type.py
@@ -7,9 +7,7 @@ from typing import Optional
 
 import numpy as np
 
-from vtsearch.config import DATA_DIR
 from vtsearch.media.base import (
-    DemoDataset,
     MediaResponse,
     MediaType,
     ProgressCallback,
@@ -103,189 +101,6 @@ class ImageMediaType(MediaType):
     @property
     def loops(self) -> bool:
         return False
-
-    # ------------------------------------------------------------------
-    # Demo datasets
-    # ------------------------------------------------------------------
-
-    _DEMO_CATEGORIES_CALTECH101 = [
-        "airplanes", "bonsai", "car_side", "chandelier", "cougar_face",
-        "crab", "dalmatian", "dolphin", "elephant", "ferry",
-        "flamingo", "grand_piano", "hawksbill", "helicopter", "ibis",
-        "kangaroo", "ketch", "lamp", "laptop", "nautilus",
-        "starfish", "stop_sign", "sunflower", "trilobite", "watch",
-    ]
-
-    _DEMO_CATEGORIES_CALTECH256 = [
-        "003.backpack", "024.butterfly", "028.camel", "029.cannon", "038.chimp",
-        "046.computer-monitor", "069.fighter-jet", "084.giraffe", "085.goat", "086.golden-gate-bridge",
-        "096.hammock", "107.hot-air-balloon", "122.kayak", "129.leopards-101", "132.light-house",
-        "145.motorbikes-101", "147.mushroom", "151.ostrich", "158.penguin", "167.pyramid",
-        "178.school-bus", "200.stained-glass", "207.swan", "245.windmill", "246.wine-bottle",
-    ]
-
-    _OXFORD_FLOWERS_CATEGORIES = [
-        "pink primrose", "hard-leaved pocket orchid", "canterbury bells", "sweet pea",
-        "english marigold", "tiger lily", "moon orchid", "bird of paradise", "monkshood",
-        "globe thistle", "snapdragon", "colt's foot", "king protea", "spear thistle",
-        "yellow iris", "globe-flower", "purple coneflower", "peruvian lily", "balloon flower",
-        "giant white arum lily", "fire lily", "pincushion flower", "fritillary", "red ginger",
-        "grape hyacinth", "corn poppy", "prince of wales feathers", "stemless gentian",
-        "artichoke", "sweet william", "carnation", "garden phlox", "love in the mist",
-        "mexican aster", "alpine sea holly", "ruby-lipped cattleya", "cape flower",
-        "great masterwort", "siam tulip", "lenten rose", "barbeton daisy", "daffodil",
-        "sword lily", "poinsettia", "bolero deep blue", "wallflower", "marigold",
-        "buttercup", "oxeye daisy", "common dandelion", "petunia", "wild pansy",
-        "primula", "sunflower", "pelargonium", "bishop of llandaff", "gaura", "geranium",
-        "orange dahlia", "pink-yellow dahlia", "cautleya spicata", "japanese anemone",
-        "black-eyed susan", "silverbush", "californian poppy", "osteospermum",
-        "spring crocus", "bearded iris", "windflower", "tree poppy", "gazania", "azalea",
-        "water lily", "rose", "thorn apple", "morning glory", "passion flower", "lotus",
-        "toad lily", "anthurium", "frangipani", "clematis", "hibiscus", "columbine",
-        "desert-rose", "tree mallow", "magnolia", "cyclamen", "watercress", "canna lily",
-        "hippeastrum", "bee balm", "ball moss", "foxglove", "bougainvillea", "camellia",
-        "mallow", "mexican petunia", "bromelia", "blanket flower", "trumpet creeper",
-        "blackberry lily",
-    ]
-
-    _FOOD101_CATEGORIES = [
-        "apple_pie", "baby_back_ribs", "baklava", "beef_carpaccio", "beef_tartare",
-        "beet_salad", "beignets", "bibimbap", "bread_pudding", "breakfast_burrito",
-        "bruschetta", "caesar_salad", "cannoli", "caprese_salad", "carrot_cake",
-        "ceviche", "cheesecake", "cheese_plate", "chicken_curry", "chicken_quesadilla",
-        "chicken_wings", "chocolate_cake", "chocolate_mousse", "churros", "clam_chowder",
-        "club_sandwich", "crab_cakes", "creme_brulee", "croque_madame", "cup_cakes",
-        "deviled_eggs", "donuts", "dumplings", "edamame", "eggs_benedict",
-        "escargots", "falafel", "filet_mignon", "fish_and_chips", "foie_gras",
-        "french_fries", "french_onion_soup", "french_toast", "fried_calamari", "fried_rice",
-        "frozen_yogurt", "garlic_bread", "gnocchi", "greek_salad", "grilled_cheese_sandwich",
-        "grilled_salmon", "guacamole", "gyoza", "hamburger", "hot_and_sour_soup",
-        "hot_dog", "huevos_rancheros", "hummus", "ice_cream", "lasagna",
-        "lobster_bisque", "lobster_roll_sandwich", "macaroni_and_cheese", "macarons",
-        "miso_soup", "mussels", "nachos", "omelette", "onion_rings", "oysters",
-        "pad_thai", "paella", "pancakes", "panna_cotta", "peking_duck", "pho",
-        "pizza", "pork_chop", "poutine", "prime_rib", "pulled_pork_sandwich",
-        "ramen", "ravioli", "red_velvet_cake", "risotto", "samosa", "sashimi",
-        "scallops", "seaweed_salad", "shrimp_and_grits", "spaghetti_bolognese",
-        "spaghetti_carbonara", "spring_rolls", "steak", "strawberry_shortcake",
-        "sushi", "tacos", "takoyaki", "tiramisu", "tuna_tartare", "waffles",
-    ]
-
-    _EUROSAT_CATEGORIES = [
-        "AnnualCrop", "Forest", "HerbaceousVegetation", "Highway", "Industrial",
-        "Pasture", "PermanentCrop", "Residential", "River", "SeaLake",
-    ]
-
-    _STANFORD_DOGS_CATEGORIES = [
-        "Chihuahua", "Japanese_spaniel", "Maltese_dog", "Pekinese", "Shih-Tzu",
-        "Blenheim_spaniel", "papillon", "toy_terrier", "Rhodesian_ridgeback", "Afghan_hound",
-        "basset", "beagle", "bloodhound", "bluetick", "black-and-tan_coonhound",
-        "Walker_hound", "English_foxhound", "redbone", "borzoi", "Irish_wolfhound",
-        "Italian_greyhound", "whippet", "Ibizan_hound", "Norwegian_elkhound", "otterhound",
-        "Saluki", "Scottish_deerhound", "Weimaraner", "Staffordshire_bullterrier",
-        "American_Staffordshire_terrier", "Bedlington_terrier", "Border_terrier",
-        "Kerry_blue_terrier", "Irish_terrier", "Norfolk_terrier", "Norwich_terrier",
-        "Yorkshire_terrier", "wire-haired_fox_terrier", "Lakeland_terrier", "Sealyham_terrier",
-        "Airedale", "cairn", "Australian_terrier", "Dandie_Dinmont", "Boston_bull",
-        "miniature_schnauzer", "giant_schnauzer", "standard_schnauzer", "Scotch_terrier",
-        "Tibetan_terrier", "silky_terrier", "soft-coated_wheaten_terrier",
-        "West_Highland_white_terrier", "Lhasa", "flat-coated_retriever", "curly-coated_retriever",
-        "golden_retriever", "Labrador_retriever", "Chesapeake_Bay_retriever",
-        "German_short-haired_pointer", "vizsla", "English_setter", "Irish_setter",
-        "Gordon_setter", "Brittany_spaniel", "clumber", "English_springer",
-        "Welsh_springer_spaniel", "cocker_spaniel", "Sussex_spaniel", "Irish_water_spaniel",
-        "kuvasz", "schipperke", "groenendael", "malinois", "briard", "kelpie", "komondor",
-        "Old_English_sheepdog", "Shetland_sheepdog", "collie", "Border_collie",
-        "Bouvier_des_Flandres", "Rottweiler", "German_shepherd", "Doberman",
-        "miniature_pinscher", "Greater_Swiss_Mountain_dog", "Bernese_mountain_dog",
-        "Appenzeller", "EntleBucher", "boxer", "bull_mastiff", "Tibetan_mastiff",
-        "French_bulldog", "Great_Dane", "Saint_Bernard", "Eskimo_dog", "malamute",
-        "Siberian_husky", "affenpinscher", "basenji", "pug", "Leonberg", "Newfoundland",
-        "Great_Pyrenees", "Samoyed", "Pomeranian", "chow", "keeshond",
-        "Brabancon_griffon", "Pembroke", "Cardigan", "toy_poodle", "miniature_poodle",
-        "standard_poodle", "Mexican_hairless", "dingo", "dhole", "African_hunting_dog",
-    ]
-
-    _UCSF_DOCUMENTS_CATEGORIES = [
-        "Tobacco", "Food", "Drug", "Chemical", "Fossil Fuel", "Opioids",
-    ]
-
-    @property
-    def demo_datasets(self) -> list:
-        from vtsearch.datasets.downloader import (  # noqa: PLC0415
-            CALTECH101_DOWNLOAD_SIZE_MB,
-            CALTECH256_DOWNLOAD_SIZE_MB,
-            EUROSAT_DOWNLOAD_SIZE_MB,
-            FOOD101_DOWNLOAD_SIZE_MB,
-            OXFORD_FLOWERS_DOWNLOAD_SIZE_MB,
-            STANFORD_DOGS_DOWNLOAD_SIZE_MB,
-            UCSF_IDL_DOWNLOAD_SIZE_MB,
-        )
-
-        cats101 = self._DEMO_CATEGORIES_CALTECH101
-        cats256 = self._DEMO_CATEGORIES_CALTECH256
-        caltech101_folder = DATA_DIR / "caltech-101" / "101_ObjectCategories"
-        caltech256_folder = DATA_DIR / "caltech-256" / "256_ObjectCategories"
-        oxford_flowers_folder = DATA_DIR / "oxford_flowers"
-        food101_folder = DATA_DIR / "food-101" / "images"
-        eurosat_folder = DATA_DIR / "EuroSAT_RGB"
-        stanford_dogs_folder = DATA_DIR / "stanford_dogs" / "Images"
-        ucsf_documents_folder = DATA_DIR / "ucsf_documents"
-        return [
-            DemoDataset(
-                id="caltech101_s", label="Caltech-101 (S)",
-                description="Centered, well-lit object photos — a classic vision benchmark.",
-                categories=cats101, source="caltech101", required_folder=caltech101_folder,
-                slice_start=0, slice_end=20, download_size_mb=CALTECH101_DOWNLOAD_SIZE_MB,
-            ),
-            DemoDataset(
-                id="caltech101_m", label="Caltech-101 (M)",
-                description="Centered, well-lit object photos — a classic vision benchmark.",
-                categories=cats101, source="caltech101", required_folder=caltech101_folder,
-                slice_start=20, slice_end=60, download_size_mb=CALTECH101_DOWNLOAD_SIZE_MB,
-            ),
-            DemoDataset(
-                id="caltech256_l", label="Caltech-256 (L)",
-                description="Harder object photos with more varied, cluttered backgrounds than Caltech-101.",
-                categories=cats256, source="caltech256", required_folder=caltech256_folder,
-                slice_start=0, slice_end=80, download_size_mb=CALTECH256_DOWNLOAD_SIZE_MB,
-            ),
-            DemoDataset(
-                id="oxford_flowers_102_a", label="Oxford Flowers 102 (A)",
-                description="Close-up flower photography with fine-grained species variation.",
-                categories=self._OXFORD_FLOWERS_CATEGORIES, source="oxford_flowers_102",
-                required_folder=oxford_flowers_folder,
-                slice_start=0, slice_end=80, download_size_mb=OXFORD_FLOWERS_DOWNLOAD_SIZE_MB,
-            ),
-            DemoDataset(
-                id="food101_a", label="Food-101 (A)",
-                description="Crowd-sourced food photos, some mislabeled — a deliberately noisy benchmark.",
-                categories=self._FOOD101_CATEGORIES, source="food101",
-                required_folder=food101_folder,
-                slice_start=0, slice_end=1000, download_size_mb=FOOD101_DOWNLOAD_SIZE_MB,
-            ),
-            DemoDataset(
-                id="eurosat_a", label="EuroSAT (A)",
-                description="Sentinel-2 satellite imagery classified by land use type.",
-                categories=self._EUROSAT_CATEGORIES, source="eurosat",
-                required_folder=eurosat_folder,
-                slice_start=0, slice_end=2700, download_size_mb=EUROSAT_DOWNLOAD_SIZE_MB,
-            ),
-            DemoDataset(
-                id="stanford_dogs_a", label="Stanford Dogs (A)",
-                description="Fine-grained dog breed photos — many visually similar breeds.",
-                categories=self._STANFORD_DOGS_CATEGORIES, source="stanford_dogs",
-                required_folder=stanford_dogs_folder,
-                slice_start=0, slice_end=171, download_size_mb=STANFORD_DOGS_DOWNLOAD_SIZE_MB,
-            ),
-            DemoDataset(
-                id="ucsf_documents_a", label="UCSF Documents (A)",
-                description="Scanned industry document pages from the UCSF Industry Documents Library.",
-                categories=self._UCSF_DOCUMENTS_CATEGORIES, source="ucsf_documents",
-                required_folder=ucsf_documents_folder,
-                slice_start=0, slice_end=25, download_size_mb=UCSF_IDL_DOWNLOAD_SIZE_MB,
-            ),
-        ]
 
     # ------------------------------------------------------------------
     # Demo dataset loading
@@ -384,7 +199,9 @@ class ImageMediaType(MediaType):
             from vtsearch.datasets.loader import load_oxford_flowers_metadata  # noqa: PLC0415
 
             flowers_dir = download_oxford_flowers(on_progress=on_progress)
-            metadata = load_oxford_flowers_metadata(flowers_dir, self._OXFORD_FLOWERS_CATEGORIES)
+            from vtsearch.datasets.importers.demo.datasets import OXFORD_FLOWERS_CATEGORIES  # noqa: PLC0415
+
+            metadata = load_oxford_flowers_metadata(flowers_dir, OXFORD_FLOWERS_CATEGORIES)
 
             by_cat: dict[str, list[tuple[Path, str]]] = {}
             for _fname, meta in sorted(metadata.items()):
